backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
pydantic-settings>=2.4.0

torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12
Pillow>=10.1.0
numpy>=1.24.0

python-dotenv>=1.0.1
mysql-connector-python>=8.0.33
sqlalchemy>=2.0.0
